Rename useSpaceActionsSetImage.ts -> useFileActionsSetImage.ts
... to reflect it's actually a file action
