+++ packages/web-app-files/src/composables/actions/files/index.ts
@@ -13,6 +13,7 @@ export * from './useFileActionsNavigate'
 export * from './useFileActionsPaste'
 export * from './useFileActionsRename'
 export * from './useFileActionsRestore'
+export * from './useFileActionsSetImage'
 export * from './useFileActionsShowActions'
 export * from './useFileActionsShowDetails'
 export * from './useFileActionsShowEditTags'
