+++ packages/web-app-files/src/composables/actions/files/useFileActionsSetImage.ts
