+++ packages/web-app-files/src/composables/actions/spaces/index.ts
@@ -1,2 +1,1 @@
-export * from './useSpaceActionsSetImage'
 export * from './useSpaceActionsUploadImage'
